Added filesystem checks to validate collection structure.
